Adapt rendering of components
